feat: remove redundant validation
